Refactor io.scala_api -> just plain "io"
