+++ src/main/scala/com/meteorcode/pathway/io/FileHandle.scala
@@ -1,4 +1,4 @@
-package com.meteorcode.pathway.io.scala_api
+package com.meteorcode.pathway.io
 
 import java.io.{
   BufferedInputStream,
@@ -10,7 +10,7 @@ import java.io.{
   }
 import java.nio.charset.Charset
 
-import com.meteorcode.pathway.io.scala_api.ResourceManager
+import com.meteorcode.pathway.io.ResourceManager
 
 import scala.io.Source
 import scala.util.{Try, Success, Failure}
