+++ src/main/scala/com/meteorcode/pathway/io/RedirectFileHandle.scala
@@ -1,7 +1,6 @@
-package com.meteorcode.pathway.io.scala_api
+package com.meteorcode.pathway.io
 
-import java.io.{File, IOException, InputStream, OutputStream}
-import java.util
+import java.io.{File, InputStream, OutputStream}
 
 import scala.util.Try
