+++ src/test/scala/com/meteorcode/pathway/test/IOSpec.scala
@@ -4,8 +4,9 @@ import java.io._
 import java.nio.charset.Charset
 import java.nio.file.{FileSystems, Files}
 
+import com.meteorcode.pathway.io.ResourceManager
 import com.meteorcode.pathway.io.java_api.AlphabeticLoadPolicy
-import com.meteorcode.pathway.io.scala_api.{ResourceManager, FilesystemFileHandle, FileHandle}
+import com.meteorcode.pathway.io.scala_api.{FilesystemFileHandle, FileHandle}
 import com.meteorcode.pathway.test.tags.FilesystemTest
 
 import org.mockito.Mockito._
